folder restructuring -validation and captcha showing is buggy
