+++ front-end/src/pages/gallery/index.js
@@ -1,25 +1,25 @@
 import ImageList from '@mui/material/ImageList';
 import ImageListItem from '@mui/material/ImageListItem';
 import { Box } from '@mui/material';
-import { Theme } from "../theme";
+import { Theme } from "../../theme";
 import { ThemeProvider } from "@mui/material/styles";
 import IconButton from '@mui/material/IconButton';
 import InfoIcon from '@mui/icons-material/Info';
 import ImageListItemBar from '@mui/material/ImageListItemBar';
 import React, { useState, useEffect } from 'react';
 
-import "./Gallery.css";
+import "./style.css";
 
-import IMG_01 from "../images/gallery/IMG_01.jpg";
-import IMG_02 from "../images/gallery/IMG_02.jpg";
-import IMG_03 from "../images/gallery/IMG_03.jpg";
-import IMG_04 from "../images/gallery/IMG_04.jpg";
-import IMG_05 from "../images/gallery/IMG_05.jpg";
-import IMG_06 from "../images/gallery/IMG_06.jpg";
-import IMG_07 from "../images/gallery/IMG_07.jpg";
-import IMG_08 from "../images/gallery/IMG_08.jpg";
-import IMG_09 from "../images/gallery/IMG_09.jpg";
-import IMG_10 from "../images/gallery/IMG_10.jpg";
+import IMG_02 from "../../images/gallery/IMG_02.jpg";
+import IMG_01 from "../../images/gallery/IMG_01.jpg";
+import IMG_03 from "../../images/gallery/IMG_03.jpg";
+import IMG_04 from "../../images/gallery/IMG_04.jpg";
+import IMG_05 from "../../images/gallery/IMG_05.jpg";
+import IMG_06 from "../../images/gallery/IMG_06.jpg";
+import IMG_07 from "../../images/gallery/IMG_07.jpg";
+import IMG_08 from "../../images/gallery/IMG_08.jpg";
+import IMG_09 from "../../images/gallery/IMG_09.jpg";
+import IMG_10 from "../../images/gallery/IMG_10.jpg";
 
 const itemData = [
   {
