+++ front-end/src/pages/gallery/style.css
