+++ front-end/src/pages/offer/style.css
